refactored account dashboard to have reusable overview box component on the left
